memisahkan views dari controller dan menambahkan modal untuk edit dan tambah
