+++ admin/models/modelpembayaran.php
@@ -1,19 +1,18 @@
 <?php
 
-    require_once 'connection.php';
+    require_once __DIR__ . '/connection.php';
+    require_once __DIR__ . '/model.php';
 
     class ModelPembayaran{
         protected $id;
         protected $nama;
         protected $keterangan;
 
-
         function __construct($id, $nama, $keterangan){
             $this->id = $id;
             $this->nama = $nama;
             $this->keterangan = $keterangan;
         }
-
 
         public static function from_array($array){
             return new ModelPembayaran(
@@ -23,36 +22,31 @@
             );
         }
 
-
         function get_id(){
             return $this->id;
         }
-
         
         function get_nama(){
             return $this->nama;
         }
 
-
         function get_keterangan(){
             return $this->keterangan;
         }
-
 
         public static function get($id){
             global $conn;
             $modelpembayaran = null;
 
-            $result -> $conn->query("SELECT * FROM model_pembayaran WHERE id_model=$id");
+            $result = $conn->query("SELECT * FROM model_pembayaran WHERE id_model=$id");
             if ($result->num_rows > 0){
-                $rawdata = $result->fetch_all();
+                $rawdata = $result->fetch_all()[0];
 
                 $modelpembayaran = ModelPembayaran::from_array($rawdata);
             }
 
             return $modelpembayaran;
         }
-
 
         public static function get_all(){
             global $conn;
@@ -69,7 +63,6 @@
 
             return $modelpembayarans;
         }
-
 
     }
