+++ admin/models/transaksi.php
@@ -1,6 +1,7 @@
 <?php
 
-    require_once "connection.php";
+    require_once __DIR__ . '/connection.php';
+    require_once __DIR__ . '/model.php';
 
     class Transaksi{
         protected $id;
